feat: add flag to set snakemake rerun triggers
